Revert to simple tool reconciliation and combine test files
- Reverted tool validation changes to use simple tool override approach
- Combined test_llm_reconciliation.py and test_local_conversation_tools_integration.py
  into single test_agent_reconciliation.py file
- Updated test expectations to match new behavior where different tools
  are allowed (they get overridden from runtime agent)
- All tests pass with the simplified reconciliation logic
